unsquashfs: fix printf format specifier for rlim_t
rlim_t can be 32 or 64 bits long (see getrlimit(2)), so cast it to
uintmax_t and use "%ju" format specifier.

+++ src/squashfs/unsquashfs.c
@@ -43,6 +43,7 @@
 #include <sys/resource.h>
 #include <limits.h>
 #include <ctype.h>
+#include <stdint.h>
 
 struct cache *fragment_cache, *data_cache;
 struct queue *to_reader, *to_inflate, *to_writer, *from_writer;
@@ -2024,8 +2025,8 @@ void initialise_threads(int fragment_buffer_size, int data_buffer_size) {
 		uint margin = (max_files - opened_files) / 10;
 		max_files -= margin;
 
-		printf("Max Files: %u, Opened Files: %u, Allocated: %u (Margin: %u)\n",
-			rlim.rlim_cur, opened_files, max_files, margin
+		printf("Max Files: %ju, Opened Files: %u, Allocated: %u (Margin: %u)\n",
+			(uintmax_t) rlim.rlim_cur, opened_files, max_files, margin
 		);
 		free(fds);
